@@ -13,16 +13,16 @@
   },
   "dependencies": {
     "@date-io/dayjs": "^2.15.0",
-    "@emotion/react": "^11.10.0",
-    "@emotion/styled": "^11.10.0",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
     "@fortawesome/fontawesome-svg-core": "^6.2.0",
     "@fortawesome/free-brands-svg-icons": "^6.2.0",
     "@fortawesome/free-regular-svg-icons": "^6.2.0",
     "@fortawesome/free-solid-svg-icons": "^6.2.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@mui/icons-material": "^5.8.4",
-    "@mui/material": "^5.9.3",
-    "@mui/styles": "^5.9.3",
+    "@mui/icons-material": "^5.15.18",
+    "@mui/material": "^5.15.18",
+    "@mui/system": "^5.15.18",
     "@mui/x-date-pickers": "^5.0.0-beta.5",
     "@react-chess/chessground": "^1.3.3",
     "chess.js": "1.0.0-alpha.0",
@@ -33,9 +33,9 @@
     "humanize-duration": "^3.27.2",
     "lodash.get": "^4.4.2",
     "lodash.uniqby": "^4.7.0",
-    "react": "^17.0.0",
+    "react": "^18.3.1",
     "react-beautiful-dnd": "^13.1.1",
-    "react-dom": "^17.0.0",
+    "react-dom": "^18.3.1",
     "react-linkify": "^1.0.0-alpha",
     "react-mentions": "^4.4.7",
     "react-resizable": "^3.0.4",
@@ -50,9 +50,9 @@
     "@types/lodash.get": "^4.4.7",
     "@types/lodash.uniqby": "^4.7.7",
     "@types/node": "^16.11.47",
-    "@types/react": "^17.0.0",
+    "@types/react": "^18.3.3",
     "@types/react-beautiful-dnd": "^13.1.5",
-    "@types/react-dom": "^17.0.0",
+    "@types/react-dom": "^18.3.0",
     "@types/react-linkify": "^1.0.1",
     "@types/react-mentions": "^4.1.6",
     "@types/react-resizable": "^3.0.3",
