@@ -3,102 +3,100 @@
  * https://codepen.io/nzbin/pen/GGrXbp
  */
 
-import { makeStyles } from '@mui/styles';
-import { Box, Theme } from '@mui/material';
+import { Box, useTheme, keyframes } from '@mui/material';
 
-const useStyles = makeStyles((theme: Theme) => ({
-  dotPulse: {
-    position: 'relative',
-    left: '-9999px',
-    width: '10px',
-    height: '10px',
-    borderRadius: '5px',
-    backgroundColor: theme.palette.primary.main,
-    color: theme.palette.primary.main,
-    boxShadow: '9999px 0 0 -5px',
-    animation: '$dot-pulse 1.5s infinite linear',
-    animationDelay: '0.25s',
-    '&::before': {
-      content: '""',
-      display: 'inline-block',
-      position: 'absolute',
-      top: '0',
-      width: '10px',
-      height: '10px',
-      borderRadius: '5px',
-      backgroundColor: theme.palette.primary.main,
-      color: theme.palette.primary.main,
+const dotPulseBefore = keyframes`
+  0% {
+    box-shadow: 9984px 0 0 -5px;
+  },
+  30% {
+    box-shadow: 9984px 0 0 2px;
+  },
+  60% {
+    box-shadow: 9984px 0 0 -5px;
+  },
+  100% {
+    box-shadow: 9984px 0 0 -5px;
+  },
+`;
 
-      boxShadow: '9984px 0 0 -5px',
-      animation: '$dot-pulse-before 1.5s infinite linear',
-      animationDelay: '0s',
-    },
-    '&::after': {
-      content: '""',
-      display: 'inline-block',
-      position: 'absolute',
-      top: '0',
-      width: '10px',
-      height: '10px',
-      borderRadius: '5px',
-      backgroundColor: theme.palette.primary.main,
-      color: theme.palette.primary.main,
+const dotPulse = keyframes`
+  0% {
+    box-shadow: 9999px 0 0 -5px;
+  },
+  30% {
+    box-shadow: 9999px 0 0 2px;
+  },
+  60% {
+    box-shadow: 9999px 0 0 -5px;
+  },
+  100% {
+    box-shadow: 9999px 0 0 -5px;
+  },
+`;
 
-      boxShadow: '10014px 0 0 -5px',
-      animation: '$dot-pulse-after 1.5s infinite linear',
-      animationDelay: '0.5s',
-    },
+const dotPulseAfter = keyframes`
+  0% {
+    box-shadow: 10014px 0 0 -5px;
   },
-
-  '@keyframes dot-pulse-before': {
-    '0%': {
-      boxShadow: '9984px 0 0 -5px',
-    },
-    '30%': {
-      boxShadow: '9984px 0 0 2px',
-    },
-    '60%': {
-      boxShadow: '9984px 0 0 -5px',
-    },
-    '100%': {
-      boxShadow: '9984px 0 0 -5px',
-    },
+  30% {
+    box-shadow: 10014px 0 0 2px;
   },
-  '@keyframes dot-pulse': {
-    '0%': {
-      boxShadow: '9999px 0 0 -5px',
-    },
-    '30%': {
-      boxShadow: '9999px 0 0 2px',
-    },
-    '60%': {
-      boxShadow: '9999px 0 0 -5px',
-    },
-    '100%': {
-      boxShadow: '9999px 0 0 -5px',
-    },
+  60% {
+    box-shadow: 10014px 0 0 -5px;
   },
-  '@keyframes dot-pulse-after': {
-    '0%': {
-      boxShadow: '10014px 0 0 -5px',
-    },
-    '30%': {
-      boxShadow: '10014px 0 0 2px',
-    },
-    '60%': {
-      boxShadow: '10014px 0 0 -5px',
-    },
-    '100%': {
-      boxShadow: '10014px 0 0 -5px',
-    },
+  100% {
+    box-shadow: 10014px 0 0 -5px;
   },
-}));
+`;
 
 export default function DotPulse(): JSX.Element {
-  const classes = useStyles();
+  const theme = useTheme();
   return (
     <Box width="45px" height="15px" display="flex" alignItems="center" justifyContent="center">
-      <div className={classes.dotPulse} />
+      <Box sx={{
+        position: 'relative',
+        left: '-9999px',
+        width: '10px',
+        height: '10px',
+        borderRadius: '5px',
+        backgroundColor: theme.palette.primary.main,
+        color: theme.palette.primary.main,
+        boxShadow: '9999px 0 0 -5px',
+        animation: `${dotPulse} 1.5s infinite linear`,
+        animationDelay: '0.25s',
+        '&::before': {
+          content: '""',
+          display: 'inline-block',
+          position: 'absolute',
+          top: '0',
+          width: '10px',
+          height: '10px',
+          borderRadius: '5px',
+          backgroundColor: theme.palette.primary.main,
+          color: theme.palette.primary.main,
+
+          boxShadow: '9984px 0 0 -5px',
+          animation: `${dotPulseBefore} 1.5s infinite linear`,
+          animationDelay: '0s',
+        },
+        '&::after': {
+          content: '""',
+          display: 'inline-block',
+          position: 'absolute',
+          top: '0',
+          width: '10px',
+          height: '10px',
+          borderRadius: '5px',
+          backgroundColor: theme.palette.primary.main,
+          color: theme.palette.primary.main,
+
+          boxShadow: '10014px 0 0 -5px',
+          animation: `${dotPulseAfter} 1.5s infinite linear`,
+          animationDelay: '0.5s',
+        },
+      }}
+      />
     </Box>
   );
 }
